Move graph API call to a separate module
